% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdxTrees_parks.R
\name{get_pdxTrees_parks}
\alias{get_pdxTrees_parks}
\title{Load the pdxTrees_parks data}
\usage{
get_pdxTrees_parks()
}
\description{
This is a function to pull the pdxTrees_parks data from github.
}
