% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdxTrees_streets.R
\name{get_pdxTrees_streets}
\alias{get_pdxTrees_streets}
\title{Load the pdxTrees_streets data}
\usage{
get_pdxTrees_streets()
}
\description{
Load the pdxTrees_streets data
}
\details{
This is a function to pull the pdxTrees_streets data from github.
}
